+++ CHANGELOG.md
@@ -2,11 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.12.5]
+
+_release `2024-11-12`_
+
+### ✨ Added
+
+- Added 'get_batches' method, a feature to retrieve a group of batches based on 'BatchFilters'
+
+### Changed
+
+- Now passing the wrong type for Filtering Jobs or Batches will raise a TypeError instead of a ValueError
+
 ## [0.12.4]
 
 _released `2024-10-09`_
 
-### ✨ Added
+### Added
 
 - Deprecation warnings on init of SDK, now each version of pasqal-cloud is supported for 1 year after release.
   If your version is not supported anymore you will get a deprecation warning each time you use the SDK.
@@ -15,7 +27,7 @@ _released `2024-10-09`_
 
 _released `2024-10-02`_
 
-### ✨ Added
+### Added
 
 - Allow unauthenticated users to access public device specifications